move `withLock` to util file and remove env reset in tests
